Updated unlinks
Updated unlinks to check to see if the file exists before calling the
unlink method.

+++ modules/kohana-archive/classes/archive.php
@@ -108,7 +108,7 @@ Abstract Class Archive
 					array(':filename' => $filename));
 
 			// Remove the file
-			unlink($filename);
+			if(file_exists($filename)) unlink($filename);
 		}
 
 		return $this->create($this->paths, $filename);
